+++ js/util.js
@@ -1,0 +1,14 @@
+// Функция взята из интернета и доработана
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+
+const getRandomPositiveInteger = (minNumber, maxNumber) => {
+  const lower = Math.ceil(Math.min(Math.abs(minNumber), Math.abs(maxNumber)));
+  const upper = Math.floor(Math.max(Math.abs(minNumber), Math.abs(maxNumber)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomPositiveInteger, getRandomElement};
